+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.0] - 2025-09-09
+
+#### Added
+- support for scarb `2.12.0`
+
 #### Removed
 
 - default `cairo-coverage` command. You must now use `cairo-coverage run` to generate a coverage report.